+++ kubejs/client_scripts/fc_villager_trades_tooltips_andre.js
@@ -1,0 +1,24 @@
+ItemEvents.tooltip(event => {
+  event.add(
+    [
+      'minecraft:saddle', 
+      'minecraft:spyglass', 
+      'farmingforblockheads:chicken_nest', 
+      'farmingforblockheads:feeding_trough', 
+      'waystones:waystone', 
+      'waystones:sandy_waystone', 
+      'waystones:mossy_waystone', 
+      'waystones:warp_stone', 
+      'mobcapturingtool:mob_capturing_tool', 
+      'naturescompass:naturescompass', 
+      'exoticbirds:bird_book', 
+      'exoticbirds:egg_identifier', 
+      'exoticbirds:egg_incubator', 
+      'exoticbirds:roost_box', 
+      'exoticbirds:pigeon_backpack'
+    ],
+    [
+      'You can buy this item from Andre'
+    ])
+
+})