+++ kubejs/client_scripts/fc_villager_trades_tooltips_laly.js
@@ -1,0 +1,26 @@
+ItemEvents.tooltip(event => {
+  event.add(
+    [
+      'minecraft:iron_ingot', 
+      'minecraft:netherite_scrap'
+    ],
+    [
+      'You can buy this item from Laly'
+    ])
+  event.add(
+    [
+      'gemsnjewels:dusk_emerald', 
+      'gemsnjewels:opal', 
+      'gemsnjewels:pale_diamond', 
+      'gemsnjewels:ruby', 
+      'gemsnjewels:sable_amethyst', 
+      'gemsnjewels:sapphire', 
+      'minecraft:redstone', 
+      'minecraft:raw_copper', 
+      'create:raw_zinc'
+    ],
+    [
+      'You can sell this item to Laly'
+    ])
+
+})