+++ kubejs/client_scripts/fc_villager_trades_tooltips_pamela.js
@@ -1,0 +1,18 @@
+ItemEvents.tooltip(event => {
+  event.add(
+    [
+      'minecraft:egg', 
+      'bakery:yeast', 
+      'minecraft:honey_bottle', 
+      'bakery:jar', 
+      'vinery:wine_bottle', 
+      'bakery:tray', 
+      'farmingforblockheads:green_fertilizer', 
+      'farmingforblockheads:red_fertilizer', 
+      'farmingforblockheads:yellow_fertilizer'
+    ],
+    [
+      'You can buy this item from Pamela'
+    ])
+
+})