+++ kubejs/client_scripts/fc_villager_trades_tooltips_sam.js
@@ -1,0 +1,34 @@
+ItemEvents.tooltip(event => {
+  event.add(
+    [
+      'pamhc2foodextended:sushiitem', 
+      'aquaculture:sushi', 
+      'pamhc2foodextended:californiarollitem', 
+      'farmersdelight:salmon_roll', 
+      'farmersdelight:cod_roll', 
+      'farmersdelight:kelp_roll_slice', 
+      'aquaculturedelight:raw_fish_fillet_roll', 
+      'moa_decor_cookery:sushi', 
+      'farmersdelight:rice_roll_medley_block', 
+      'aquaculture:algae', 
+      'aquaculture:lockbox', 
+      'aquaculture:treasure_chest', 
+      'minecraft:trident', 
+      'aquaculture:neptunium_ingot'
+    ],
+    [
+      'You can buy this item from Sam'
+    ])
+  event.add(
+    [
+      'aquaculture:catfish', 
+      'aquaculture:capitaine', 
+      'aquaculture:tuna', 
+      'aquaculture:arrau_turtle', 
+      'aquaculture:starshell_turtle'
+    ],
+    [
+      'You can sell this item to Sam'
+    ])
+
+})